services/ai_model/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
joblib
pandas
numpy
google-cloud-storage
psycopg2-binary
sqlalchemy
pydantic-settings
